Fix data type coin

+++ app/src/main/java/com/example/playcoin/data/data_source/dto/coinDetailDTO/CodeAdditionsDeletions4Weeks.kt
@@ -1,6 +1,6 @@
 package com.example.playcoin.data.data_source.dto.coinDetailDTO
 
 data class CodeAdditionsDeletions4Weeks(
-    val additions: Int,
-    val deletions: Int
+    val additions: Double,
+    val deletions: Double
 )
