+++ app/src/main/java/com/example/playcoin/data/data_source/dto/coinDetailDTO/CommunityData.kt
@@ -2,10 +2,10 @@ package com.example.playcoin.data.data_source.dto.coinDetailDTO
 
 data class CommunityData(
     val facebook_likes: Any,
-    val reddit_accounts_active_48h: Int,
+    val reddit_accounts_active_48h: Double,
     val reddit_average_comments_48h: Double,
     val reddit_average_posts_48h: Double,
-    val reddit_subscribers: Int,
+    val reddit_subscribers: Double,
     val telegram_channel_user_count: Any,
-    val twitter_followers: Int
+    val twitter_followers: Double
 )
