+++ app/src/main/java/com/example/playcoin/data/data_source/dto/coinDetailDTO/PriceChangePercentage14dInCurrency.kt
@@ -11,7 +11,7 @@ data class PriceChangePercentage14dInCurrency(
     val bmd: Double,
     val bnb: Double,
     val brl: Double,
-    val btc: Int,
+    val btc: Double,
     val cad: Double,
     val chf: Double,
     val clp: Double,
